Move waitUntil to it's own file
