+++ src/utils.ts
@@ -3,13 +3,11 @@ import type { ParsedCSSValue } from 'webdriverio'
 
 import { expect } from 'expect'
 
-import { DEFAULT_OPTIONS } from './constants.js'
 import type { WdioElementMaybePromise } from './types.js'
 import { wrapExpectedWithArray } from './util/elementsUtil.js'
 import { executeCommand } from './util/executeCommand.js'
 import { enhanceError, enhanceErrorBe } from './util/formatMessage.js'
-
-const sleep = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms))
+import { waitUntil } from './util/waitUntil.js'
 
 export function isJasmineStringAsymmetricMatcher<T>(expected: unknown): expected is JasmineAsymmetricMatcher<T> {
     return isAsymmetricMatcher(expected) && 'expected' in expected
@@ -63,59 +61,6 @@ export function getAsymmetricMatcherValue<T>(
 
     // Jasmine anything, truthy, falsy, empty, notEmpty asymmetric matchers do not have a sample or expected value. So cannot throw an error here. Return undefined to indicate that there is no value to extract.
     return undefined
-}
-
-/**
- * wait for expectation to succeed
- * @param condition function
- * @param isNot     https://jestjs.io/docs/expect#thisisnot
- * @param options   wait, interval, etc
- */
-const waitUntil = async (
-    condition: () => Promise<boolean>,
-    isNot = false,
-    { wait = DEFAULT_OPTIONS.wait, interval = DEFAULT_OPTIONS.interval } = {}
-): Promise<boolean> => {
-    // single attempt
-    if (wait === 0) {
-        return await condition()
-    }
-
-    let error: Error | undefined
-
-    // wait for condition to be truthy
-    try {
-        const start = Date.now()
-        while (true) {
-            if (Date.now() - start > wait) {
-                throw new Error('timeout')
-            }
-
-            try {
-                const result = isNot !== (await condition())
-                error = undefined
-                if (result) {
-                    break
-                }
-                await sleep(interval)
-            } catch (err) {
-                error = err
-                await sleep(interval)
-            }
-        }
-
-        if (error) {
-            throw error
-        }
-
-        return !isNot
-    } catch {
-        if (error) {
-            throw error
-        }
-
-        return isNot
-    }
 }
 
 async function executeCommandBe(
